LayoutTests/imported/w3c/web-platform-tests/html/the-xhtml-syntax/parsing-xhtml-documents/xml-stylesheet-blocking-expected.txt


FAIL xml-stylesheet blocks script execution and rendering assert_equals: XML processing instruction should've blocked script execution and rendering expected "3" but got "auto"
